Refactoring state objects to use a common item class
